feat: remove keep or cancel ci form (#32)
