+++ action-server/covidflow/actions/action_qa_goodbye.py
@@ -4,7 +4,7 @@ from rasa_sdk import Action, Tracker
 from rasa_sdk.events import ConversationPaused
 from rasa_sdk.executor import CollectingDispatcher
 
-from covidflow.constants import CONTINUE_CI_SLOT
+from covidflow.constants import CANCEL_CI_SLOT, LAST_SYMPTOMS_SLOT
 
 from .lib.log_util import bind_logger
 
@@ -20,7 +20,11 @@ class ActionQaGoodbye(Action):
         domain: Dict[Text, Any],
     ) -> List[Dict[Text, Any]]:
         bind_logger(tracker)
-        if tracker.get_slot(CONTINUE_CI_SLOT) is True:
+        # Last symptoms slot tells us we are in a daily check-in conversation
+        if (
+            tracker.get_slot(LAST_SYMPTOMS_SLOT) is not None
+            and tracker.get_slot(CANCEL_CI_SLOT) is not True
+        ):
             dispatcher.utter_message(
                 template="utter_daily_ci__qa__will_contact_tomorrow"
             )
