+++ action-server/tests/actions/test_action_qa_goodbye.py
@@ -2,7 +2,7 @@ import pytest
 from rasa_sdk.events import ConversationPaused
 
 from covidflow.actions.action_qa_goodbye import ActionQaGoodbye
-from covidflow.constants import CONTINUE_CI_SLOT
+from covidflow.constants import CANCEL_CI_SLOT, LAST_SYMPTOMS_SLOT, Symptoms
 
 from .action_test_helper import ActionTestCase
 
@@ -13,8 +13,8 @@ class ActionQAGoodbyeTest(ActionTestCase):
         self.action = ActionQaGoodbye()
 
     @pytest.mark.asyncio
-    async def test_continue_ci_true(self):
-        tracker = self.create_tracker(slots={CONTINUE_CI_SLOT: True})
+    async def test_continue_ci(self):
+        tracker = self.create_tracker(slots={LAST_SYMPTOMS_SLOT: Symptoms.MILD})
 
         await self.run_action(tracker)
 
@@ -25,8 +25,20 @@ class ActionQAGoodbyeTest(ActionTestCase):
         )
 
     @pytest.mark.asyncio
-    async def test_continue_ci_false(self):
-        tracker = self.create_tracker(slots={CONTINUE_CI_SLOT: False})
+    async def test_cancel_ci(self):
+        tracker = self.create_tracker(
+            slots={LAST_SYMPTOMS_SLOT: Symptoms.MILD, CANCEL_CI_SLOT: True}
+        )
+
+        await self.run_action(tracker)
+
+        self.assert_events([ConversationPaused()])
+
+        self.assert_templates(["utter_goodbye"])
+
+    @pytest.mark.asyncio
+    async def test_not_in_ci(self):
+        tracker = self.create_tracker()
 
         await self.run_action(tracker)
