+++ core/domain/domain.core.yml
@@ -54,7 +54,7 @@ actions:
   - action_check_task_allowed
   - action_visit_package
   - action_initialize_daily_checkin
-  - action_daily_ci_early_opt_out
+  - action_daily_ci_cancel_ci
   - action_explain_preconditions
   - action_goodbye
   - action_greeting_messages
@@ -90,7 +90,6 @@ actions:
   - action_ask_daily_ci_enroll__validation_code
   - action_ask_preconditions
   - action_daily_ci_enroll_form_ended
-
   - action_submit_daily_ci_assessment
   - validate_daily_ci_feel_better_form
   - action_ask_daily_ci_feel_better_form_has_other_mild_symptoms
@@ -98,10 +97,11 @@ actions:
   - validate_daily_ci_feel_worse_form
   - action_ask_daily_ci_feel_worse_form_has_cough
   - action_ask_daily_ci_feel_worse_form_has_diff_breathing
-  - action_set_feel_worse_true
   - validate_daily_ci_feel_no_change_form
   - action_ask_daily_ci_feel_no_change_form_has_cough
   - action_ask_daily_ci_feel_no_change_form_has_fever
+  - action_check_mandatory_ci
+  - action_keep_ci_recommendations
 
 forms:
   - home_assistance_form:
@@ -162,7 +162,6 @@ forms:
       has_diff_breathing: *boolean
       daily_ci_feel_worse_form_has_diff_breathing_worsened: *boolean
       has_cough: *boolean
-  - daily_ci_keep_or_cancel_form:
   - question_answering_form:
       active_question:
         - type: from_text
@@ -300,10 +299,7 @@ slots:
   has_assistance:
     type: unfeaturized
 
-  feel_worse:
-    type: unfeaturized
-
-  continue_ci:
+  cancel_ci:
     type: unfeaturized
 
   provincial_811:
@@ -322,6 +318,9 @@ slots:
     type: unfeaturized
 
   invalid_reminder_id:
+    type: bool
+
+  mandatory_ci:
     type: bool
 
   checkin_return__moderate_symptoms_worsened:
